created error handling

+++ src/users/users.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { ConflictException, Injectable, NotFoundException } from '@nestjs/common';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
 import { UsersRepository } from './repository/users.repository';
@@ -10,6 +10,11 @@ export class UsersService {
   constructor(private readonly repository: UsersRepository){}
 
   async create({ name, email, password }: CreateUserDto): Promise<UserEntity> {
+    const emailExists = await this.repository.emailExists(email)
+
+    if(emailExists) {
+      throw new ConflictException('Previous existence of the user with the same email!')
+    }
     const hashedPassword = await hash(password, 10)
     return this.repository.create({
       name,
@@ -18,19 +23,35 @@ export class UsersService {
     });
   }
 
-  findAll(): Promise<UserEntity[]> {
+  async findAll(): Promise<UserEntity[]> {
     return this.repository.findAll();
   }
 
-  findOne(id: string): Promise<UserEntity> {
-    return this.repository.findOne(id);
+  async findOne(id: string): Promise<UserEntity> {
+    const userExists = await this.repository.findOne(id);
+    if(!userExists) {
+      throw new NotFoundException('User not found!')
+    }
+    return userExists
   }
 
-  update(id: string, updateUserDto: UpdateUserDto): Promise<UserEntity> {
+  async update(id: string, updateUserDto: UpdateUserDto): Promise<UserEntity> {
+    const userExists = await this.repository.findOne(id)
+
+    if(!userExists) {
+      throw new NotFoundException('User not found!')
+    }
+
     return this.repository.update(id, updateUserDto);
   }
 
-  remove(id: string): Promise<UserEntity> {
+  async remove(id: string): Promise<UserEntity> {
+    const userExists = await this.repository.findOne(id)
+
+    if(!userExists) {
+      throw new NotFoundException('User not found!')
+    }
+
     return this.repository.remove(id);
   }
 }
